requirements.txt
Flask
pytest
psycopg2-binary